String will be a decimal

+++ packages/beacon-node/src/api/impl/validator/index.ts
@@ -463,7 +463,10 @@ export function getValidatorApi(
 
       metrics?.blockProductionSuccess.inc({source});
       metrics?.blockProductionNumAggregated.observe({source}, block.body.attestations.length);
-      metrics?.blockProductionExecutionPayloadValue.observe({source}, parseInt(formatWeiToEth(executionPayloadValue)));
+      metrics?.blockProductionExecutionPayloadValue.observe(
+        {source},
+        parseFloat(formatWeiToEth(executionPayloadValue))
+      );
       logger.verbose("Produced blinded block", {
         slot,
         executionPayloadValue,
@@ -538,7 +541,10 @@ export function getValidatorApi(
 
       metrics?.blockProductionSuccess.inc({source});
       metrics?.blockProductionNumAggregated.observe({source}, block.body.attestations.length);
-      metrics?.blockProductionExecutionPayloadValue.observe({source}, parseInt(formatWeiToEth(executionPayloadValue)));
+      metrics?.blockProductionExecutionPayloadValue.observe(
+        {source},
+        parseFloat(formatWeiToEth(executionPayloadValue))
+      );
       logger.verbose("Produced execution block", {
         slot,
         executionPayloadValue,
